# api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Smart home api",
    "description": "This documentation describes APIs found under https://github.com/e154/smart-home",
    "version": "1.0",
    "contact": {
      "name": "Alex Filippov",
      "url": "https://e154.github.io/smart-home/",
      "email": "support@e154.ru"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/user": {
      "post": {
        "summary": "add new user",
        "operationId": "UserService_AddUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserFull"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiNewtUserRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/user/{id}": {
      "get": {
        "summary": "get user by id",
        "operationId": "UserService_GetUserById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserFull"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "user id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "summary": "delete user by id",
        "operationId": "UserService_DeleteUserById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "user id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "put": {
        "summary": "update user by id",
        "operationId": "UserService_UpdateUserById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserFull"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nickname": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "passwordRepeat": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "lang": {
                  "type": "string"
                },
                "image": {
                  "$ref": "#/definitions/apiUpdateUserRequestImage"
                },
                "role": {
                  "$ref": "#/definitions/apiUpdateUserRequestRole"
                },
                "meta": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/apiUserMeta"
                  }
                }
              }
            }
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/users": {
      "get": {
        "summary": "get user list",
        "operationId": "UserService_GetUserList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetUserListResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiGetUserListResult": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiUserShot"
          }
        },
        "meta": {
          "$ref": "#/definitions/apiMeta"
        }
      }
    },
    "apiImage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "thumb": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "mimeType": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiMeta": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "string",
          "format": "uint64"
        },
        "page": {
          "type": "string",
          "format": "uint64"
        },
        "total": {
          "type": "string",
          "format": "uint64"
        },
        "sort": {
          "type": "string"
        }
      }
    },
...
    "apiNewtUserRequest": {
      "type": "object",
      "properties": {
        "nickname": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "passwordRepeat": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "lang": {
          "type": "string"
        },
        "image": {
          "$ref": "#/definitions/apiNewtUserRequestImage"
        },
        "role": {
          "$ref": "#/definitions/apiNewtUserRequestRole"
        },
        "meta": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiUserMeta"
          }
        }
      }
    },
    "apiNewtUserRequestImage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiNewtUserRequestRole": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "apiRole": {
      "type": "object",
      "properties": {
        "parent": {
          "$ref": "#/definitions/apiRole"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiRole"
          }
        },
        "accessList": {
          "$ref": "#/definitions/apiRoleAccessList"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiRoleAccessList": {
      "type": "object",
      "properties": {
        "levels": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/AccessListListOfString"
          }
        }
      }
    },
...
    "apiUpdateUserRequestImage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiUpdateUserRequestRole": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "apiUserFull": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "nickname": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "history": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiUserHistory"
          }
        },
        "image": {
          "$ref": "#/definitions/apiImage"
        },
        "signInCount": {
          "type": "string",
          "format": "int64"
        },
        "meta": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiUserMeta"
          }
        },
        "role": {
          "$ref": "#/definitions/apiRole"
        },
        "roleName": {
          "type": "string"
        },
        "lang": {
          "type": "string"
        },
        "authenticationToken": {
          "type": "string"
        },
        "currentSignInIp": {
          "type": "string"
        },
        "lastSignInIp": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/apiUserFullParent"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "currentSignInAt": {
          "type": "string",
          "format": "date-time"
        },
        "lastSignInAt": {
          "type": "string",
          "format": "date-time"
        },
        "resetPasswordSentAt": {
          "type": "string",
          "format": "date-time"
        },
        "deletedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiUserFullParent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "nickname": {
          "type": "string"
        }
      }
    },
    "apiUserHistory": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string"
        },
        "time": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiUserMeta": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "x-api-key",
      "in": "header",
      "x-api-key": "jwt"
    },
    "BasicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "ApiKeyAuth": [],
      "BasicAuth": []
    }
  ]
}